#82 Moved configuration classes to separate config folder
